--- a/babel.config.js
+++ b/babel.config.js
@@ -1,7 +1,6 @@
 module.exports = {
-	presets: ['@babel/preset-react', '@babel/preset-env'],
+	presets: ['@babel/preset-env'],
 	plugins: [
-		'@babel/plugin-proposal-object-rest-spread',
 		'@babel/plugin-syntax-dynamic-import',
 	],
 	env: {
@@ -16,7 +15,6 @@ module.exports = {
 			],
 			plugins: [
 				'@babel/plugin-transform-runtime',
-				'@babel/plugin-proposal-class-properties',
 			],
 		},
 		test: {
@@ -32,7 +30,6 @@ module.exports = {
 			],
 			plugins: [
 				'@babel/plugin-transform-runtime',
-				'@babel/plugin-proposal-class-properties',
 				'dynamic-import-node',
 			],
 		},