--- a/jest.config.js
+++ b/jest.config.js
@@ -5,14 +5,10 @@ const esModules = ['@guardian/', 'lodash-es'].join('|');
 
 module.exports = {
 	clearMocks: true,
-	preset: 'ts-jest/presets/js-with-babel',
-	moduleFileExtensions: ['ts', 'tsx', 'js'],
 	testMatch: ['**/*.(test|spec).+(ts|tsx|js)'],
 	roots: ['<rootDir>/src'],
 	moduleDirectories: ['<rootDir>/src', 'node_modules'],
 	moduleNameMapper: {
-		videojs: 'video.js',
-		raven: '<rootDir>/src/lib/__mocks__/raven.ts',
 		'^svgs/(.*)$': '<rootDir>/src/__mocks__/svgMock.js',
 		'^(.*)\\.svg$': '<rootDir>/src/__mocks__/svgMock.js',
 		'^(.*)\\.html$': '<rootDir>/src/__mocks__/templateMock.js',
@@ -24,6 +20,6 @@ module.exports = {
 	},
 	transformIgnorePatterns: [`/node_modules/.pnpm/(?!${esModules})`],
 	transform: {
-		'^.+\\.(ts|tsx)$': 'ts-jest',
+		'^.+\\.(ts|tsx|js)$': '@swc/jest',
 	},
 };