--- a/package.json
+++ b/package.json
@@ -40,10 +40,6 @@
 		"validate": "npm-run-all tsc lint test build"
 	},
 	"devDependencies": {
-		"@babel/cli": "^7.26.4",
-		"@babel/core": "^7.26.0",
-		"@babel/plugin-proposal-class-properties": "^7.18.6",
-		"@babel/plugin-proposal-object-rest-spread": "^7.20.7",
 		"@babel/plugin-syntax-dynamic-import": "^7.8.3",
 		"@babel/plugin-transform-runtime": "^7.25.9",
 		"@babel/preset-env": "^7.26.0",
@@ -106,9 +102,11 @@
 	"dependencies": {
 		"@guardian/prebid.js": "8.52.0-10",
 		"@octokit/core": "^6.1.2",
+		"@swc/jest": "0.2.37",
 		"fastdom": "^1.0.12",
 		"lodash-es": "^4.17.21",
 		"process": "^0.11.10",
+		"swc-loader": "0.2.6",
 		"tslib": "^2.7.0",
 		"web-vitals": "^4.2.4"
 	},