--- a/package.json
+++ b/package.json
@@ -58,6 +58,7 @@
 		"@guardian/prettier": "8.0.1",
 		"@guardian/source": "8.0.1",
 		"@playwright/test": "1.49.1",
+		"@swc/jest": "0.2.37",
 		"@types/googletag": "~3.1.3",
 		"@types/jest": "^29.5.14",
 		"@types/lodash-es": "^4.17.12",
@@ -88,6 +89,7 @@
 		"npm-run-all": "^4.1.5",
 		"prettier": "^3.4.2",
 		"raw-loader": "^4.0.2",
+		"swc-loader": "0.2.6",
 		"terser-webpack-plugin": "^5.3.11",
 		"ts-jest": "^29.2.5",
 		"ts-loader": "^9.5.1",
@@ -102,11 +104,9 @@
 	"dependencies": {
 		"@guardian/prebid.js": "8.52.0-10",
 		"@octokit/core": "^6.1.2",
-		"@swc/jest": "0.2.37",
 		"fastdom": "^1.0.12",
 		"lodash-es": "^4.17.21",
 		"process": "^0.11.10",
-		"swc-loader": "0.2.6",
 		"tslib": "^2.7.0",
 		"web-vitals": "^4.2.4"
 	},