--- a/src/init/consented/prepare-prebid.ts
+++ b/src/init/consented/prepare-prebid.ts
@@ -35,7 +35,7 @@ const prebidVersion = () => {
 const loadPrebid = async (framework: ConsentFramework): Promise<void> => {
 	if (shouldLoadPrebid()) {
 		await import(
-			/* webpackChunkName: "Prebid.js" */
+			/* webpackChunkName: "[request]" */
 			`@guardian/prebid.js/build/dist/${prebidVersion()}/prebid`
 		);
 		prebid.initialise(window, framework);