--- a/webpack.config.prod.mjs
+++ b/webpack.config.prod.mjs
@@ -1,6 +1,5 @@
 import { execSync } from 'child_process';
 import { join } from 'path';
-import TerserPlugin from 'terser-webpack-plugin';
 import webpack from 'webpack';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import { merge } from 'webpack-merge';
@@ -43,8 +42,4 @@ export default merge(config, {
 			...gitCommitSHA(),
 		}),
 	],
-	optimization: {
-		minimize: true,
-		minimizer: [new TerserPlugin()],
-	},
 });