--- a/webpack.config.prod.mjs
+++ b/webpack.config.prod.mjs
@@ -54,7 +54,6 @@ export default merge(config, {
 					},
 					compress: true,
 				},
-
 			}),
 		],
 	},