--- a/webpack.config.riff-raff.mjs
+++ b/webpack.config.riff-raff.mjs
@@ -1,5 +1,6 @@
 import { execSync } from 'child_process';
 import { join } from 'path';
+import TerserPlugin from 'terser-webpack-plugin';
 import webpack from 'webpack';
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 import { merge } from 'webpack-merge';
@@ -111,4 +112,17 @@ export default merge(config, {
 		}),
 		new GenerateCloudformation(),
 	],
+	optimization: {
+		minimize: true,
+		minimizer: [
+			new TerserPlugin({
+				minify: TerserPlugin.swcMinify,
+				terserOptions: {
+					mangle: true,
+					compress: true,
+				},
+
+			}),
+		],
+	},
 });