--- a/webpack.config.riff-raff.mjs
+++ b/webpack.config.riff-raff.mjs
@@ -121,7 +121,6 @@ export default merge(config, {
 					mangle: true,
 					compress: true,
 				},
-
 			}),
 		],
 	},